old_python/requirements.txt
absl-py==0.12.0
amqp==5.0.5
attrs==20.3.0
bidict==0.21.2
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
Flask==1.1.2
Flask-SocketIO==5.0.1
freetype-py==2.2.0
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
kombu==5.0.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mediapipe==0.8.3.1
networkx==2.5
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
posix-ipc==1.0.5
prompt-toolkit==3.0.18
protobuf==3.15.6
pyglet==1.5.15
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.5
pyparsing==2.4.7
pyrender==0.1.45
python-dateutil==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
redis==3.5.3
scipy==1.6.1
shared-ndarray==1.1.1
six==1.15.0
trimesh==3.9.10
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.3.0

